Fix test after refactoring
